requirements.txt
flask
boto3
qrcode
pillow
